pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "startup-auditor"
version = "0.1.0"
description = "Automated startup analysis for AI/ML engineers targeting AI-first startups"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.24.0,<0.25.0",
    "rich",
    "python-dotenv",
    "playwright",
    "httpx",
    "pytest",
]

[project.scripts]
startup-auditor = "startup_auditor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/startup_auditor"]
